requirements.txt
capstone==5.0.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
gensim==4.3.3
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
Levenshtein==0.25.1
lief==0.15.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==2.6.3
numpy==1.22.4
packaging==24.1
pandas==1.3.5
pefile==2024.8.26
pillow==10.2.0
pip==24.2
PyGSP==0.5.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
python-louvain==0.16
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
scikit-learn==1.3.2
scipy==1.10.1
setuptools==75.1.0
six==1.16.0
smart-open==7.0.5
sympy==1.12
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.9.0
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.2
